Turn the test into a unit test
